+++ projects/ngx-form-object/src/lib/decorators/relationship-form-object-builder/relationship-form-object-builder.decorator.ts
@@ -1,0 +1,10 @@
+import { FormObject } from '../../form-object/form-object';
+
+export function RelationshipFormObjectBuilder(propertyName: string): MethodDecorator {
+	return function (target: FormObject, _key: string | symbol, descriptor: PropertyDescriptor) {
+		target.createFormObjectMethods = target.createFormObjectMethods || new Map();
+		target.createFormObjectMethods.set(propertyName, descriptor.value);
+
+		return descriptor;
+	};
+}