+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### 9.0.3-beta
 
   * Temporary adjust save method types
+  * Implement RelationshipFormObjectBuilder decorator
 
 ### 9.0.2-beta